Fixed instructions for header includes and linking.
CMake's install doesn't put the igraph headers in the place where LinkingTo
expects them to be, so we have to manually set up the pkgconfig() for that.

We also need each downstream package to link to R's BLAS and LAPACK because we
that would be available promised during igraph's compilation.